Use the native workerd perf_hooks modules when available
Previously, the unenv-preset always used a polyfill for perf_hooks, which was provided in the unenv libray.
This polyfill used sniffing of the `performance.addEventListener` method to determine if the environment supported
the perf_hooks API natively, and if so, it would use the native implementation.

However, now that the native `performance` object has been updated to provide stubs for many of these methods,
the sniffing approach no longer works.

This change moves the perf_hooks polyfill into the unenv-preset package, and removes the sniffing logic.
The decision whether to use the native or polyfill implementation is now made based on whether the
`enable_nodejs_perf_hooks_module` flag is set.